refactor(framework-core): move decorators and calculations to correct folders
