+++ packages/framework/libraries/core/src/reflectMetadata/data/catchExceptionMetadataReflectKey.ts
@@ -1,0 +1,2 @@
+export const catchExceptionMetadataReflectKey: string =
+  '@inversifyjs/framework-core/exception-filter/catchExceptionMetadataReflectKey';
