+++ packages/framework/libraries/core/src/reflectMetadata/data/classExceptionFilterMetadataReflectKey.ts
@@ -1,0 +1,2 @@
+export const classExceptionFilterMetadataReflectKey: string =
+  '@inversifyjs/framework-core/class/classExceptionFilterMetadataReflectKey';
