+++ packages/framework/libraries/core/src/reflectMetadata/data/classMethodExceptionFilterMetadataReflectKey.ts
@@ -1,0 +1,2 @@
+export const classMethodExceptionFilterMetadataReflectKey: string =
+  '@inversifyjs/framework-core/class/classMethodExceptionFilterMetadataReflectKey';
